feat: add mcp gift ideas tools
