+++ app/mcp_server/tools/gift_ideas/mark_as_bought_tool.rb
@@ -1,15 +1,13 @@
 # frozen_string_literal: true
 
-module GiftersMcp
-  module Tools
-    class MarkAsBuyingTool < MCP::Tool
-      description "Marque une idée de cadeau comme 'en cours d'achat'. " \
-                  "IMPORTANT : demander une confirmation explicite à l'utilisateur avant d'appeler cet outil. " \
-                  "Si l'utilisateur a des enfants (comptes gérés), demander au nom de qui il achète (actor_id)."
+module Tools
+  module GiftIdeas
+    class MarkAsBoughtTool < MCP::Tool
+      description "Marque une idée de cadeau comme 'achetée'. " \
+                  "IMPORTANT : demander une confirmation explicite à l'utilisateur avant d'appeler cet outil."
       input_schema(
         properties: {
-          id:       { type: "integer", description: "ID de l'idée de cadeau (obtenu via search_gift_ideas)" },
-          actor_id: { type: "integer", description: "ID de l'acheteur : current_user ou un de ses enfants (optionnel, défaut = current_user)" }
+          id: { type: "integer", description: "ID de l'idée de cadeau (obtenu via search_gift_ideas)" }
         },
         required: %w[id]
       )
@@ -25,17 +23,14 @@ module GiftersMcp
       )
 
       class << self
-        def call(server_context:, id:, actor_id: nil)
+        def call(server_context:, id:)
           user = user_from_context(server_context)
           gift_idea = GiftIdea.find_by(id: id)
 
           return not_found_response unless gift_idea
-          return unauthorized_response unless GiftIdeaPolicy.new(user, gift_idea).mark_as_buying?
+          return unauthorized_response unless GiftIdeaPolicy.new(user, gift_idea).mark_as_bought?
 
-          buyer = resolve_buyer(user, actor_id)
-          return unauthorized_response unless buyer
-
-          gift_idea.mark_as_buying(buyer)
+          gift_idea.mark_as_bought(user)
           data = serialize(gift_idea)
           MCP::Tool::Response.new(
             [{ type: "text", text: data.to_json }],
@@ -50,17 +45,6 @@ module GiftersMcp
         def user_from_context(server_context)
           user_id = server_context[:user_id] || server_context["user_id"]
           User.find(user_id)
-        end
-
-        def resolve_buyer(user, actor_id)
-          return user if actor_id.nil?
-
-          actor = User.find_by(id: actor_id)
-          return nil unless actor
-          return actor if actor.id == user.id
-          return actor if user.can_access_as_parent?(actor)
-
-          nil
         end
 
         def serialize(g)
