--- lib/gifters_mcp.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-module GiftersMcp
-end
