+++ spec/mcp_server/tools/gift_ideas/mark_as_bought_tool_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe GiftersMcp::Tools::MarkAsBoughtTool do
+RSpec.describe Tools::GiftIdeas::MarkAsBoughtTool do
   let(:buyer) { create(:user) }
   let(:server_context) { { user_id: buyer.id } }
